Factor out the cloudflare spans into separate files
